Fix to upgrade only vrouter and rsyslod in ziu_compute
In ziu_compute, install_contrail will be executed after stopping vrouter.
If there is a controller definition in instances.yaml, the controller will also be upgraded.
In the above case, after stopping vrouter, controller and vrouter will be upgraded, so the downtime of the service will be longer.
So, I changed ziu_compute to upgrade only vrouter and ryslogd in compute node.

Issue-ID: TFB-1643
Signed-off-by: s.maeda <s.maeda@ntt.com>
Change-Id: Ie6ce8723ae8c2c767ada5968f3ee1b089bd1f321